refactor(ui): extract "isRedisVersionSupported" to separate util
re #RI-7363
