+++ redisinsight/ui/src/pages/vector-search/create-index/hooks/useRedisInstanceCompatibility.ts
@@ -1,7 +1,7 @@
-import semver from 'semver'
 import { useSelector } from 'react-redux'
 import { connectedInstanceSelector } from 'uiSrc/slices/instances/instances'
 import { REDISEARCH_MODULES } from 'uiSrc/slices/interfaces'
+import { isRedisVersionSupported } from 'uiSrc/utils/comparisons/compareVersions'
 
 export type UseRedisInstanceCompatibilityReturn = {
   loading: boolean | undefined
@@ -11,23 +11,6 @@ export type UseRedisInstanceCompatibilityReturn = {
 
 const MIN_SUPPORTED_REDIS_VERSION = '7.2.0'
 const REDISEARCH_MODULE_SET = new Set(REDISEARCH_MODULES)
-
-export const isVersionSupported = (raw: string): boolean => {
-  // Try a loose/full parse of the whole string first.
-  // This returns a normalized version string like "7.2.0" or null if not recognizable.
-  const vLoose = semver.valid(raw, { loose: true })
-  if (vLoose)
-    return semver.satisfies(vLoose, `>=${MIN_SUPPORTED_REDIS_VERSION}`)
-
-  // Fallback: try to coerce a version from arbitrary text,
-  // e.g. "Redis 7.2.1" -> SemVer { version: '7.2.1' }
-  const coerced = semver.coerce(raw)
-  if (!coerced) {
-    return false
-  }
-
-  return semver.gte(coerced, MIN_SUPPORTED_REDIS_VERSION)
-}
 
 const useRedisInstanceCompatibility =
   (): UseRedisInstanceCompatibilityReturn => {
@@ -45,7 +28,9 @@ const useRedisInstanceCompatibility =
       loading,
       hasRedisearch: !loading ? hasRedisearch : undefined,
       hasSupportedVersion:
-        !loading && version ? isVersionSupported(version) : undefined,
+        !loading && version
+          ? isRedisVersionSupported(version, MIN_SUPPORTED_REDIS_VERSION)
+          : undefined,
     }
   }
