+++ redisinsight/ui/src/utils/comparisons/compareVersions.ts
@@ -1,3 +1,5 @@
+import semver from 'semver'
+
 export const isVersionHigherOrEquals = (
   sourceVersion: string = '',
   comparableVersion: string = '',
@@ -42,4 +44,23 @@ export const isVersionHigher = (
   }
 
   return false
+}
+
+export const isRedisVersionSupported = (
+  raw: string,
+  minVersion: string,
+): boolean => {
+  // Try a loose/full parse of the whole string first.
+  // This returns a normalized version string like "7.2.0" or null if not recognizable.
+  const vLoose = semver.valid(raw, { loose: true })
+  if (vLoose) return semver.satisfies(vLoose, `>=${minVersion}`)
+
+  // Fallback: try to coerce a version from arbitrary text,
+  // e.g. "Redis 7.2.1" -> SemVer { version: '7.2.1' }
+  const coerced = semver.coerce(raw)
+  if (!coerced) {
+    return false
+  }
+
+  return semver.gte(coerced, minVersion)
 }
