+++ redisinsight/ui/src/utils/tests/comparisons/compareVersions.spec.ts
@@ -1,4 +1,8 @@
-import { isVersionHigherOrEquals, isVersionHigher } from 'uiSrc/utils'
+import {
+  isVersionHigherOrEquals,
+  isVersionHigher,
+  isRedisVersionSupported,
+} from 'uiSrc/utils'
 
 describe('isVersionHigherOrEqual', () => {
   it('isVersionHigherOrEqual should return true if the first version provided is higher or equal', () => {
@@ -45,5 +49,28 @@ describe('isVersionHigher', () => {
     expect(isVersionHigher(version7, '6.2')).toBeTruthy()
     expect(isVersionHigher(version8, '6.2')).toBeTruthy()
     expect(isVersionHigher(version9, '6.2')).toBeTruthy()
+  })
+})
+
+describe('useRedisInstanceCompatibility', () => {
+  const MOCK_MIN_SUPPORTED_REDIS_VERSION = '7.2.0'
+
+  test.each([
+    [undefined, false],
+    [null, false],
+    ['', false],
+    ['7.1.9', false],
+    ['7.2', true],
+    ['v7.2', true],
+    ['7.2.0', true],
+    ['7.2.0+build.5', true],
+    ['7.2.0-rc.1', false], // prerelease should NOT satisfy >=7.2.0
+    ['Redis 7.2-rc1 (abc)', true], // coerced to 7.2.0 -> true (note: prerelease info lost)
+    ['Redis 6 something', false],
+    ['nonsense', false],
+  ])('isRedisVersionSupported(%p) === %p', (input, expected) => {
+    expect(
+      isRedisVersionSupported(input as any, MOCK_MIN_SUPPORTED_REDIS_VERSION),
+    ).toBe(expected)
   })
 })
